Complete Project by adding ability to toggle read status by clicking
Used callback funciton with the click event listener to call the
toggleRead prototype on my book. in the callback function I refresh
the display of books so that the changes are visualized. I nested
the addEventListener statement in the display book function so that
it would automatically be applied to all books.